requirements.txt
Flask
supabase
python-dotenv
flask-cors
